requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.17

numpy==1.26.4
scikit-learn==1.5.2
faiss-cpu==1.9.0.post1
rank-bm25==0.2.2

PyMuPDF==1.24.14
pytesseract==0.3.13
opencv-python-headless==4.10.0.84
Pillow==11.0.0

sentence-transformers==3.3.1
transformers==4.46.3
torch>=2.4.1

langdetect==1.0.9
nltk==3.9.1
rapidfuzz==3.10.1
requests==2.32.3
orjson==3.10.12
